requirements.txt
alembic==1.7.5
aniso8601==9.0.1
asn1crypto==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
distlib==0.3.4
filelock==3.4.2
flasgger==0.9.5
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mistune==2.0.2
pg8000==1.23.0
pipenv==2022.1.8
platformdirs==2.4.1
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scramp==1.4.1
six==1.16.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
testing.common.database==2.0.3
text-unidecode==1.3
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
Werkzeug==2.0.2
